+++ pkg/ccache/statistics_test.go
@@ -1,6 +1,5 @@
-// Copyright 2018-2022 VirtualTam.
-// Use of this source code is governed by a MIT license
-// that can be found in the LICENSE file.
+// Copyright (c) VirtualTam
+// SPDX-License-Identifier: MIT
 
 package ccache