+++ cmd/ccache_exporter/metrics/server.go
@@ -38,8 +38,8 @@ func accessLogger(r *http.Request, status, size int, dur time.Duration) {
 }
 
 // NewServer registers metrics collectors and returns a HTTP server to expose them.
-func NewServer(wrapper *ccache.LocalCommand, listenAddr string, versionDetails *version.Details) *http.Server {
-	ccacheCollector := newCcacheCollector(wrapper)
+func NewServer(ccacheWrapper *ccache.Wrapper, listenAddr string, versionDetails *version.Details) *http.Server {
+	ccacheCollector := newCcacheCollector(ccacheWrapper)
 	versionCollector := newVersionCollector("ccache_exporter", versionDetails)
 
 	prometheus.MustRegister(