+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [v4.0.0](https://github.com/virtualtam/ccache_exporter/releases/tag/v4.0.0) - UNRELEASED
 
-Minor refactoring with breaking changes.
+Major refactoring with breaking changes.
 
 ### Breaking
 
@@ -20,6 +20,7 @@ Minor refactoring with breaking changes.
 - Add a contributing guide
 - Setup Hashicorp Copywrite to automate license header management
 - Allow configuring the exporter via command-line flags, environment variables and configuration file
+- Add subcommand to display the exporter version
 
 ### Changed